requirements.txt
streamlit==1.20.0
mysql-connector-python==8.0.33
pandas==2.0.3
numpy==1.24.3
requests==2.31.0
spacy==3.6.1
whisper==1.0.0
python-dotenv
PyPDF2
openai